requirements.txt
# Backend Framework
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-multipart>=0.0.12

# Cloudinary Image CDN
cloudinary>=1.44.0

# Pinecone Vector Database
pinecone>=7.0.0

# Deep Learning & Computer Vision
torch>=2.6.0
torchvision>=0.17.0
Pillow>=10.4.0

# Quantum Computing
qiskit>=1.0.2
qiskit-aer>=0.13.3

# Scientific Computing
numpy>=1.24.3
scipy>=1.11.3

# Configuration & Environment
python-dotenv>=1.0.0

# Utilities
python-json-logger>=2.0.0

# Performance & Caching
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0

# Advanced Models
transformers>=4.35.0
timm>=0.9.0

# Vector Compression
faiss-cpu>=1.7.0

# Rate Limiting
slowapi>=0.1.9

# WebSockets
python-socketio>=5.10.0

# Monitoring
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0

# Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.0
